Enable postSubBuffer on D3D11 with DXGI 1.2.
BUG=147291

Reviewed-on: https://chromium-review.googlesource.com/270572
Tested-by: Sunny Sachanandani <sunnyps@chromium.org>

+++ src/libANGLE/renderer/d3d/d3d9/SwapChain9.cpp
@@ -312,8 +312,8 @@ EGLint SwapChain9::swapRect(EGLint x, EGLint y, EGLint width, EGLint height)
 
     RECT rect =
     {
-        x, mHeight - y - height,
-        x + width, mHeight - y
+        static_cast<LONG>(x), static_cast<LONG>(mHeight - y - height),
+        static_cast<LONG>(x + width), static_cast<LONG>(mHeight - y)
     };
 
     HRESULT result = mSwapChain->Present(&rect, &rect, NULL, NULL, 0);
